Add argument for padded event times
When structuring data, PoPPy creates dictionaries for each event
containing it and its historical event times and types. In the case
that there are not enough historical events to fill the memory, e.g.,
the first event, PoPPy will randomly create events to pad out the memory
size. These events occur at the time of first event and the event type
is randomly assigned. This causes strange intensity function values to
be calculated since there are several events now occurring at the time
of the first event.

To fix this:

- Padded historical events now default to time of 0.
- Extra argument added so user can change time for padded events.